Align synchronization declarations of runtime:runtime module with other compose modules (ui:ui, ui:ui-text, foundation:foundation)
- rename SynchronizedObject.kt to Synchronization.kt
- move it to androidx.compose.runtime.platform package
