+++ compose/runtime/runtime/src/androidMain/kotlin/androidx/compose/runtime/platform/Synchronization.android.kt
@@ -14,12 +14,14 @@
  * limitations under the License.
  */
 
-package androidx.compose.runtime
+package androidx.compose.runtime.platform
 
-import kotlin.DeprecationLevel.*
+@Suppress("ACTUAL_WITHOUT_EXPECT") // https://youtrack.jetbrains.com/issue/KT-37316
+internal actual typealias SynchronizedObject = Any
+
+@Suppress("NOTHING_TO_INLINE")
+internal actual inline fun makeSynchronizedObject(ref: Any?) = ref ?: SynchronizedObject()
 
 @PublishedApi
-@JvmName("synchronized")
-@Deprecated(level = HIDDEN, message = "should not be used")
-internal inline fun <R> oldSynchronized(lock: SynchronizedObject, block: () -> R): R =
-    androidx.compose.runtime.platform.synchronized(lock, block)
+internal actual inline fun <R> synchronized(lock: SynchronizedObject, block: () -> R): R =
+    kotlin.synchronized(lock, block)
