+++ compose/runtime/runtime/src/commonMain/kotlin/androidx/compose/runtime/snapshots/SnapshotStateObserver.kt
@@ -29,11 +29,11 @@ import androidx.compose.runtime.composeRuntimeError
 import androidx.compose.runtime.internal.AtomicReference
 import androidx.compose.runtime.internal.currentThreadId
 import androidx.compose.runtime.internal.currentThreadName
-import androidx.compose.runtime.makeSynchronizedObject
+import androidx.compose.runtime.platform.makeSynchronizedObject
 import androidx.compose.runtime.observeDerivedStateRecalculations
 import androidx.compose.runtime.requirePrecondition
 import androidx.compose.runtime.structuralEqualityPolicy
-import androidx.compose.runtime.synchronized
+import androidx.compose.runtime.platform.synchronized
 
 /**
  * Helper class to efficiently observe snapshot state reads. See [observeReads] for more details.
