+++ compose/runtime/runtime/src/desktopMain/kotlin/androidx/compose/runtime/platform/Synchronization.desktop.kt
@@ -13,17 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package androidx.compose.runtime.platform
 
-package androidx.compose.runtime
+@Suppress("ACTUAL_WITHOUT_EXPECT") // https://youtrack.jetbrains.com/issue/KT-37316
+internal actual typealias SynchronizedObject = androidx.compose.runtime.SynchronizedObject
 
-internal expect class SynchronizedObject
-
-/**
- * Returns [ref] as a [SynchronizedObject] on platforms where [Any] is a valid [SynchronizedObject],
- * or a new [SynchronizedObject] instance if [ref] is null or this is not supported on the current
- * platform.
- */
-internal expect inline fun makeSynchronizedObject(ref: Any? = null): SynchronizedObject
+@Suppress("NOTHING_TO_INLINE")
+internal actual inline fun makeSynchronizedObject(ref: Any?) = SynchronizedObject()
 
 @PublishedApi
-internal expect inline fun <R> synchronized(lock: SynchronizedObject, block: () -> R): R
+internal actual inline fun <R> synchronized(lock: SynchronizedObject, block: () -> R): R =
+    kotlin.synchronized(lock, block)
