+++ compose/runtime/runtime/src/posixMain/kotlin/androidx/compose/runtime/platform/Synchronization.posix.kt
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package androidx.compose.runtime
+package androidx.compose.runtime.platform
 
 import androidx.compose.runtime.internal.currentThreadId
 import kotlin.native.ref.createCleaner
